Add files via upload
This is a simple calculator made using Python programming language. You can perform basic operations such as addition, multiplication, subtraction, division, modulus and a clear operation. This is the first task I did related to the CodSoft online internship program. 
It took me less than a day to finish this task as it was simple and even the logic was easy.