Add files via upload
This is the second task, a password generator having 3 different levels namely "easy, medium and hard". In 'easy' level there's use of just uppercase and lowercase letters. In 'medium' there's addition of digits/numbers with upper and lower case letters. Lastly, there's 'difficult' level where symbols is also added with everything else. 
After the generation of the password according to the length given by the user, there's a 'display' button, where by pressing it the password is displayed on the screen for the user to view. The user can keep pressing the 'display' button to generate new passwords for as long as they don't get the desired password or whatsoever.