Add files via upload
This is a simple to-do-list I've created with the help of a listbox widget in python gui. 
You add elements/tasks as you wish whenever you want. Once you're done with the pending task you can simply delete it from the list and move on to do the next task!